+++ CHANGELOG.rst
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+- Update penalty to shrinkage parameter conversion in synthetic likelihood calculation
+- Update BSL pre sample methods
+- Update BslSample
+- Move ROMC tests
 - Update to numpy 1.24
 - Update readthedocs configuration
 - Restrict numpy < 1.24 until codebase has been updated